Hmmm my comments about 0 and nil being the same in the previous checking was just plain wrong. The problem was this:
The emacs type is only decided on compile time of the file. If you ever use
GNU Emacs, then it will compile the file, and it will say emacs. If you
then later uses XEmacs, then it will load the compiled file, and the
variable will claim that you are using GNU Emacs, not XEmacs.

The other way around is not a problem, as GNU Emacs will not load XEmacs
compiled files.

git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/KDE/kdesdk/scripts/kde-emacs@745545 283d02a7-25f6-0310-bc7c-ecb5cbfe19da